pyproject.toml
[project]
name = "midas"
version = "0.1.0"
description = "Personal finance MCP server — net worth, budgets, brokerage performance"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
]

[project.optional-dependencies]
dotenv = ["python-dotenv>=1.0.0"]
dev = ["pytest>=8.0.0", "ruff>=0.9.0"]

[project.scripts]
midas = "src.server:main"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]
